Move evaluation to strategies
